[SPARK-37841][SQL] BasicWriteTaskStatsTracker should not try get status for a skipped file
### What changes were proposed in this pull request?

In the comment https://github.com/apache/spark/pull/35117#issuecomment-1007171965 in #35117 and other details, we noticed that the BasicWriteTaskStatsTracker keeps complaining that some of the written files are missing. This is truely worrisome.

```scala
scala> spark.range(0, 3, 1, 4).write.option("compression", "zstd").orc("./spark-warehouse/zstd_1")
22/01/07 14:22:48 WARN BasicWriteTaskStatsTracker: Expected 1 files, but only saw 0. This could be due to the output format not writing empty files, or files being not immediately visible in the filesystem.
```

This happens for ORC format and the root cause is because, in OrcOutputWriter, the `recordWriter` and the `file` are lazily generated once have a record to write. So this step is skipped when the records iterator is empty. In the `commit` part, we will use the `file` to update stats assuming it exists. This causes `FileNotFoundException` and results not stats update. Then, `numSubmittedFiles` and `numFiles` become inconsistent and issue a false warning like above.

In this PR, we make the hive.OrcOutputWriter and OrcFileFormat behave the same as other data sources. That said, it will always create a file even when there are no records at all.

### Why are the changes needed?

fix a regression
- the warning is wrong
- make builtin data sources behave same

### Does this PR introduce _any_ user-facing change?

no

### How was this patch tested?

added ut

locally

```log
Welcome to
      ____              __
     / __/__  ___ _____/ /__
    _\ \/ _ \/ _ `/ __/  '_/
   /___/ .__/\_,_/_/ /_/\_\   version 3.3.0-SNAPSHOT
      /_/

Using Scala version 2.12.15 (OpenJDK 64-Bit Server VM, Java 11.0.12)
Type in expressions to have them evaluated.
Type :help for more information.

scala> spark.range(3).write.option("compression", "zstd").orc("./spark-warehouse/zstdtest_23")

scala> spark.range(3).write.option("compression", "zstd").orc("./spark-warehouse/zstdtest_24")

scala>
```

Closes #35132 from yaooqinn/SPARK-37841.

Authored-by: Kent Yao <yao@apache.org>
Signed-off-by: Kent Yao <yao@apache.org>